Libraries package rename. #tommyettinger/gdx-liftoff#60
