+++ src/main/kotlin/gdx/liftoff/data/libraries/unofficial/lml.kt
@@ -1,9 +1,9 @@
 @file:Suppress("unused") // Extension classes accessed via reflection.
 
-package gdx.liftoff.data.libs.unofficial
+package gdx.liftoff.data.libraries.unofficial
 
-import gdx.liftoff.data.libs.Repository
-import gdx.liftoff.data.libs.camelCaseToKebabCase
+import gdx.liftoff.data.libraries.Repository
+import gdx.liftoff.data.libraries.camelCaseToKebabCase
 import gdx.liftoff.data.platforms.*
 import gdx.liftoff.data.project.Project
 import gdx.liftoff.views.Extension
