+++ src/main/kotlin/gdx/liftoff/data/templates/unofficial/webSocketBasic.kt
@@ -2,8 +2,8 @@ package gdx.liftoff.data.templates.unofficial
 
 import gdx.liftoff.data.files.CopiedFile
 import gdx.liftoff.data.files.path
-import gdx.liftoff.data.libs.unofficial.VisUI
-import gdx.liftoff.data.libs.unofficial.Websocket
+import gdx.liftoff.data.libraries.unofficial.VisUI
+import gdx.liftoff.data.libraries.unofficial.Websocket
 import gdx.liftoff.data.platforms.Assets
 import gdx.liftoff.data.project.Project
 import gdx.liftoff.data.templates.Template
